Francesco: modified the doubles to be floats in the project.  About to agressively change the table format for the action value function.

+++ openrocket/core/src/net/sf/openrocket/simulation/extension/impl/RLObjectFileStore.java
@@ -35,8 +35,8 @@ public class RLObjectFileStore {
         return (ArrayList) readObjects(episodeDataFileName);
     }
 
-    public HashMap<StateActionTuple, Double> readActionValueFunction(String filenameExtension){
-        return (HashMap<StateActionTuple, Double>) readObjects(actionValueFunctionFileName + filenameExtension);
+    public HashMap<StateActionTuple, Float> readActionValueFunction(String filenameExtension){
+        return (HashMap<StateActionTuple, Float>) readObjects(actionValueFunctionFileName + filenameExtension);
     }
 
     public void storeActionValueFunction(HashMap actionValueFunction, String filenameExtension){
